pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "kosta-enhanced-word-mcp-server"
version = "2.0.0"
description = "Enhanced Word MCP server for academic research collaboration with advanced search/replace, review tools, and section management"
readme = "README.md"
license = {file = "LICENSE"}
authors = [
    {name = "Kosta Vučković", email = "kosta@brown.edu"},
    {name = "GongRzhe", email = "gongrzhe@gmail.com"}
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
requires-python = ">=3.11"
dependencies = [
    "python-docx>=0.8.11",
    "mcp[cli]>=1.3.0",
    "msoffcrypto-tool>=5.4.2",
    "docx2pdf>=0.1.8",
]

[project.urls]
"Homepage" = "https://github.com/kosta/kosta-enhanced-word-mcp-server"
"Bug Tracker" = "https://github.com/kosta/kosta-enhanced-word-mcp-server/issues"

[tool.hatch.build.targets.wheel]
only-include = [
    "word_document_server",
    "office_word_mcp_server",
]
sources = ["."]

[project.scripts]
word_mcp_server = "word_document_server.main:run_server"
